Refactor react component
